requirements.txt
django-bootstrap4==2.2.0
django-extensions==3.0.5
django-recaptcha==2.0.6
django-select2==7.4.2
Pillow==7.2.0
requests==2.24.0

